\name{BayesMVProbit-package}
\alias{BayesMVProbit-package}
\alias{BayesMVProbit}
\docType{package}
\title{
\packageTitle{BayesMVProbit}
}
\description{
\packageDescription{BayesMVProbit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesMVProbit}
\packageIndices{BayesMVProbit}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BayesMVProbit}

Maintainer: \packageMaintainer{BayesMVProbit}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
