% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate Nominal.r
\name{nominal_post_beta}
\alias{nominal_post_beta}
\title{Title}
\usage{
nominal_post_beta(
  category,
  iter,
  burn,
  cred_level = 0.95,
  x_list,
  sig,
  d,
  prior_beta_mean = NULL,
  prior_beta_var = NULL
)
}
\arguments{
\item{category}{}

\item{iter}{}

\item{burn}{}

\item{cred_level}{}

\item{x_list}{}

\item{sig}{}

\item{d}{}

\item{prior_beta_mean}{}

\item{prior_beta_var}{}
}
\value{

}
\description{
Title
}
