% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate Ordinal.r
\name{ordinal_post_beta}
\alias{ordinal_post_beta}
\title{Title}
\usage{
ordinal_post_beta(
  category,
  df_t = NULL,
  iter,
  burn,
  cred_level = 0.95,
  x_list,
  sig,
  y,
  prior_delta_mean = NULL,
  prior_delta_var = NULL,
  prior_beta_mean = NULL,
  prior_beta_var = NULL
)
}
\arguments{
\item{category}{}

\item{df_t}{}

\item{iter}{}

\item{burn}{}

\item{cred_level}{}

\item{x_list}{}

\item{sig}{}

\item{y}{}

\item{prior_delta_mean}{}

\item{prior_delta_var}{}

\item{prior_beta_mean}{}

\item{prior_beta_var}{}
}
\value{

}
\description{
Title
}
